The blinking algorithm is pretty much done.
It uses both A & B channels of Timer0 in Fast PWM mode. Both channels are connected to their Output Compare pins. The Timer0 Overflow Interrupt powers a state machine, which in turn toggles the DDR bit for both Output Compare pins.

The Timer0 controls the brightness of the LEDs with PWM, and the state machine controls the blinking pattern.